fix checkpointing (#541)

+++ open_instruct/utils.py
@@ -666,7 +666,7 @@ def get_last_checkpoint(folder: str, incomplete: bool = False) -> Optional[str]:
         checkpoints = [path for path in checkpoints if os.path.exists(os.path.join(folder, path, "COMPLETED"))]
     if len(checkpoints) == 0:
         return
-    return os.path.join(folder, max(checkpoints, key=lambda x: x.split("_")[-1]))
+    return os.path.join(folder, max(checkpoints, key=lambda x: int(x.split("_")[-1])))
 
 
 def get_last_checkpoint_path(args, incomplete: bool = False) -> str:
@@ -693,7 +693,7 @@ def clean_last_n_checkpoints(output_dir: str, keep_last_n_checkpoints: int) -> N
     folders = [f for f in os.listdir(output_dir) if is_checkpoint_folder(output_dir, f)]
     # find the checkpoint with the largest step
     checkpoints = sorted(folders, key=lambda x: int(x.split("_")[-1]))
-    if len(checkpoints) > keep_last_n_checkpoints:
+    if keep_last_n_checkpoints >= 0 and len(checkpoints) > keep_last_n_checkpoints:
         for checkpoint in checkpoints[: len(checkpoints) - keep_last_n_checkpoints]:
             logger.info(f"Removing checkpoint {checkpoint}")
             shutil.rmtree(os.path.join(output_dir, checkpoint))
